several updates; main changes: inclusion of automated html report; bam files are allowed as input
git-svn-id: https://hedgehog.fhcrc.org/bioconductor/trunk/madman/Rpacks/TEQC@57804 bc3139a8-67e5-0310-9ffc-ced21a209358

+++ R/coverage.GC.R
@@ -27,7 +27,8 @@ function(coverageAll, baits, returnBaitValues=FALSE, linecol="darkred", lwd, xla
   }
 
   # average coverage for bait-covered bases
-  avgcov <- mean(as.numeric(unlist(covercounts.baits)))
+#  avgcov <- mean(as.numeric(unlist(covercounts.baits)))
+  avgcov <- mean(as.integer(unlist(covercounts.baits)))
 
   # normalized per-bait coverage
   baitcov.norm <- baitcov / avgcov
